moved the .descr-details styles and scripts into its own 'example' block for usage outside of the accordion
